backend/requirements.txt
accelerate
addict
basicsr==1.4.2
bitsandbytes
controlnet_aux
diffusers
einops
facexlib>=0.2.5
fastapi
future
gfpgan>=1.3.5
h5py
imageio
kornia
lmdb
lpips
mediapipe
netron
ninja
numpy==1.24.1
onnx
onnxruntime
open_clip_torch
opencv-python==4.7.0.72
packaging
Pillow==9.4.0
protobuf==3.20.3
psutil
pyyaml
python-multipart
requests
safetensors
scikit-image
scipy
setuptools==65.5.0
tb-nightly
thop
torch>=1.7
torchvision
tqdm
transformers
uvicorn
xformers
yapf